requirements.txt
streamlit>=1.32.0
tensorflow>=2.15.0
pillow>=10.1.0
numpy>=1.24.3
pandas>=2.1.4  